Move single fetch shared types into RR
